requirements.txt
aiogram
APScheduler
art
colorama
psutil
requests
snscrape
SQLAlchemy
ujson
python-dotenv
yt-dlp
aiosqlite
